# backend/docs/swagger.json
{
    "schemes": [
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "Backend service for SmartLeague. Cookie-based auth (HttpOnly access/refresh cookies).",
        "title": "SmartLeague API",
        "contact": {
            "name": "SmartLeague Backend"
        },
        "version": "1.0"
    },
    "host": "localhost:8000",
    "paths": {
        "/api/v1/auth/refresh": {
            "post": {
                "description": "Refreshes access and refresh tokens using valid refresh token from cookies",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "token"
                ],
                "summary": "Refresh tokens",
                "responses": {
                    "204": {
                        "description": "OK"
                    },
                    "401": {
                        "description": "Unauthorized - invalid/missing refresh token",
                        "schema": {
                            "$ref": "#/definitions/dto.HTTPStatus"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/dto.HTTPStatus"
                        }
                    }
                }
            }
        },
...
        "/api/v1/series/{id}": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "series"
                ],
                "summary": "Get series by id",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Series ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.GetSeriesResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/dto.HTTPStatus"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/dto.HTTPStatus"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "CookieAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "series"
                ],
                "summary": "Delete series by id",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Series ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/dto.HTTPStatus"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/dto.HTTPStatus"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/dto.HTTPStatus"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/dto.HTTPStatus"
                        }
                    }
                }
            },
            "patch": {
                "security": [
                    {
                        "CookieAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "series"
                ],
                "summary": "Update series by id",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Series ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Update data",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.UpdateSeriesRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.UpdateSeriesResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/dto.HTTPStatus"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/dto.HTTPStatus"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/dto.HTTPStatus"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/dto.HTTPStatus"
                        }
                    }
                }
            }
        },
...
        "/api/v1/user/all": {
            "get": {
                "description": "Retrieves a list of users filtered by role, one-line full name query, and email prefix.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "user"
                ],
                "summary": "Get users by filters",
                "parameters": [
                    {
                        "maximum": 100,
                        "minimum": 1,
                        "type": "integer",
                        "example": 10,
                        "description": "Max number of users to return",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "minimum": 0,
                        "type": "integer",
                        "example": 0,
                        "description": "Pagination offset",
                        "name": "offset",
                        "in": "query"
                    },
                    {
                        "enum": [
                            0,
                            1,
                            2
                        ],
                        "type": "integer",
                        "example": 0,
                        "description": "Role enum (0–2)",
                        "name": "role",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "example": "\"Иван Дима\"",
                        "description": "One-line search by name, nickname and email tokens",
                        "name": "q",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "example": "\"user@\"",
                        "description": "Filter by email prefix",
                        "name": "email_prefix",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.GetAllByFilterUsersResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid query parameters",
                        "schema": {
                            "$ref": "#/definitions/dto.HTTPStatus"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/dto.HTTPStatus"
                        }
                    }
                }
            }
        },
...
        "/api/v1/user/password": {
            "post": {
                "security": [
                    {
                        "CookieAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "user"
                ],
                "summary": "Change user password",
                "parameters": [
                    {
                        "description": "Password change data",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.ChangePasswordRequest"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content",
                        "headers": {
                            "Set-Cookie": {
                                "type": "string",
                                "description": "user_auth_refresh_token=token; Path=/; HttpOnly; Secure; SameSite=Strict"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/dto.HTTPStatus"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/dto.HTTPStatus"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/dto.HTTPStatus"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/dto.HTTPStatus"
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "$ref": "#/definitions/dto.HTTPStatus"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/dto.HTTPStatus"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "dto.ChangePasswordRequest": {
            "type": "object",
            "required": [
                "new_password",
                "old_password"
            ],
            "properties": {
                "new_password": {
                    "type": "string",
                    "format": "password",
                    "maxLength": 100,
                    "minLength": 8,
                    "example": "NewPass456!"
                },
                "old_password": {
                    "type": "string",
                    "format": "password",
                    "maxLength": 100,
                    "minLength": 8,
                    "example": "OldPass123!"
                }
            }
        },
...
        "dto.GetAllByFilterUsersResponse": {
            "type": "object",
            "properties": {
                "items": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dto.User"
                    }
                },
                "pagination": {
                    "$ref": "#/definitions/dto.PaginationInfo"
                }
            }
        },
...
        "dto.GetSeriesResponse": {
            "type": "object",
            "properties": {
                "club_id": {
                    "type": "string"
                },
                "creator_id": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "end_at": {
                    "type": "string"
                },
                "game_type": {
                    "$ref": "#/definitions/types.GameType"
                },
                "id": {
                    "type": "string"
                },
                "is_closed": {
                    "type": "boolean"
                },
                "name": {
                    "type": "string"
                },
                "price_rub": {
                    "type": "integer"
                },
                "start_at": {
                    "type": "string"
                },
                "status": {
                    "$ref": "#/definitions/types.SeriesStatus"
                }
            }
        },
...
        "dto.HTTPStatus": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "message": {
                    "type": "string"
                }
            }
        },
...
        "dto.PaginationInfo": {
            "type": "object",
            "properties": {
                "current_page": {
                    "type": "integer",
                    "example": 1
                },
                "has_next": {
                    "type": "boolean",
                    "example": true
                },
                "has_previous": {
                    "type": "boolean",
                    "example": false
                },
                "total_items": {
                    "type": "integer",
                    "example": 1250
                },
                "total_pages": {
                    "type": "integer",
                    "example": 13
                }
            }
        },
...
        "dto.UpdateSeriesRequest": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string",
                    "maxLength": 10000,
                    "minLength": 1
                },
                "end_at": {
                    "type": "string"
                },
                "game_type": {
                    "$ref": "#/definitions/types.GameType"
                },
                "is_closed": {
                    "type": "boolean"
                },
                "name": {
                    "type": "string",
                    "maxLength": 200,
                    "minLength": 1
                },
                "price_rub": {
                    "type": "integer",
                    "maximum": 100000000,
                    "minimum": 0
                },
                "start_at": {
                    "type": "string"
                },
                "status": {
                    "maximum": 3,
                    "minimum": 0,
                    "allOf": [
                        {
                            "$ref": "#/definitions/types.SeriesStatus"
                        }
                    ]
                }
            }
        },
        "dto.UpdateSeriesResponse": {
            "type": "object",
            "properties": {
                "club_id": {
                    "type": "string"
                },
                "creator_id": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "end_at": {
                    "type": "string"
                },
                "game_type": {
                    "$ref": "#/definitions/types.GameType"
                },
                "id": {
                    "type": "string"
                },
                "is_closed": {
                    "type": "boolean"
                },
                "name": {
                    "type": "string"
                },
                "price_rub": {
                    "type": "integer"
                },
                "start_at": {
                    "type": "string"
                },
                "status": {
                    "$ref": "#/definitions/types.SeriesStatus"
                }
            }
        },
...
        "dto.User": {
            "type": "object",
            "properties": {
                "club_id": {
                    "type": "string",
                    "example": "550e8400-e29b-41d4-a716-446655440000"
                },
                "club_state": {
                    "allOf": [
                        {
                            "$ref": "#/definitions/types.ClubState"
                        }
                    ],
                    "example": 0
                },
                "description": {
                    "type": "string",
                    "example": "About me"
                },
                "email": {
                    "type": "string",
                    "example": "user@example.com"
                },
                "id": {
                    "type": "string",
                    "example": "550e8400-e29b-41d4-a716-446655440000"
                },
                "name": {
                    "type": "string",
                    "example": "Ivan"
                },
                "nickname": {
                    "type": "string",
                    "example": "mishmish"
                },
                "role": {
                    "allOf": [
                        {
                            "$ref": "#/definitions/types.Role"
                        }
                    ],
                    "example": 0
                },
                "show_name": {
                    "type": "boolean",
                    "example": true
                }
            }
        },
        "types.ClubState": {
            "type": "integer",
            "enum": [
                0,
                1,
                2,
                3,
                4
            ],
            "x-enum-varnames": [
                "ClubStateNone",
                "ClubStateMember",
                "ClubStateLeader",
                "ClubStatePresident",
                "ClubStateResident"
            ]
        },
...
        "types.GameType": {
            "type": "integer",
            "enum": [
                0
            ],
            "x-enum-varnames": [
                "GameTypeSportMafia"
            ]
        },
...
        "types.Role": {
            "type": "integer",
            "enum": [
                0,
                1,
                2,
                3
            ],
            "x-enum-varnames": [
                "RoleUser",
                "RoleModerator",
                "RoleAdmin",
                "RoleSuperAdmin"
            ]
        },
        "types.SeriesStatus": {
            "type": "integer",
            "enum": [
                0,
                1,
                2,
                3
            ],
            "x-enum-varnames": [
                "SeriesStatusClosed",
                "SeriesStatusRegistration",
                "SeriesStatusClosedRegistration",
                "SeriesStatusGames"
            ]
        }
    },
    "securityDefinitions": {
        "CookieAuth": {
            "description": "Authentication via HttpOnly cookies:\\n- `user_auth_access_token` (short-lived)\\n- `user_auth_refresh_token` (long-lived)\\n\\nProtected endpoints require valid cookies to be sent by the client.",
            "type": "apiKey",
            "name": "user_auth_access_token",
            "in": "cookie"
        }
    },
...
}